make executable dependent tests optional locally
