+++ tox.ini
@@ -103,7 +103,7 @@ commands = python {toxinidir}/bin/rebuild-support.py
 [testenv:fix_lint]
 description = format the code base to adhere to our styles, and complain about what we cannot do automatically
 basepython = python3.7
-passenv = HOMEPATH PROGRAMDATA http_proxy https_proxy no_proxy
+passenv = *
 deps = pre-commit == 1.12.0
 skip_install = True
 commands = pre-commit run --all-files --show-diff-on-failure
